@@ -11,29 +11,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.1.3",
-    "@angular/common": "~10.1.3",
-    "@angular/compiler": "~10.1.3",
-    "@angular/core": "~10.1.3",
-    "@angular/forms": "~10.1.3",
-    "@angular/platform-browser": "~10.1.3",
-    "@angular/platform-browser-dynamic": "~10.1.3",
-    "@angular/router": "~10.1.3",
+    "@angular/animations": "~11.2.9",
+    "@angular/common": "~11.2.9",
+    "@angular/compiler": "~11.2.9",
+    "@angular/core": "~11.2.9",
+    "@angular/forms": "~11.2.9",
+    "@angular/platform-browser": "~11.2.9",
+    "@angular/platform-browser-dynamic": "~11.2.9",
+    "@angular/router": "~11.2.9",
+    "@zxing/browser": "0.0.8",
+    "@zxing/library": "^0.18.4",
+    "@zxing/ngx-scanner": "^3.1.3",
+    "angular2-qrcode": "^2.0.3",
+    "angularx-qrcode": "^11.0.0",
     "rxjs": "~6.6.0",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1001.7",
-    "@angular/cli": "~10.1.3",
-    "@angular/compiler-cli": "~10.1.3",
-    "@types/jasmine": "~3.5.0",
+    "@angular-devkit/build-angular": "^0.1102.8",
+    "@angular/cli": "~11.2.8",
+    "@angular/compiler-cli": "~11.2.9",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
