@@ -23,6 +23,9 @@ import { CategoryItemComponent } from './category-item/category-item.component';
 import { CategoryListPageComponent } from './category-list-page/category-list-page.component';
 import { SharePageComponent } from './share-page/share-page.component'
 
+import { QRCodeModule } from 'angularx-qrcode';
+import { ZXingScannerModule } from '@zxing/ngx-scanner';
+
 
 
 @NgModule({
@@ -49,7 +52,9 @@ import { SharePageComponent } from './share-page/share-page.component'
     AppRoutingModule,
     CommonModule,
     BrowserModule,
-    HttpClientModule
+    HttpClientModule,
+    QRCodeModule,
+    ZXingScannerModule
     
    
   ],
